refactor: move and rename TextBlockCard and PrtableTextULTick
