+++ src/components/UnitAndLessonLists/UnitList/UnitListItem/UnitListItem.tsx
@@ -65,8 +65,8 @@ const UnitListItem: FC<UnitListItemProps> = (props) => {
       subjectIconBackground === "pink"
         ? "pink60"
         : subjectIconBackground === "lavender"
-          ? "lavender60"
-          : subjectIconBackground;
+        ? "lavender60"
+        : subjectIconBackground;
   }
 
   return (
